Source Code Frequencia app Python
This repository contains the source code for a simple attendance tracking system developed for Giro Ferragista. The system is built using Flask and Python, allowing employees to log their entry and exit times automatically. It uses CSV files to store user and attendance data. This project was developed as part of an extension activity to streamline employee attendance management in small businesses.